Reorganized some Svg2Ico code
 * Moved some classes into new files
 * Nested some locally used classes
 * Changed visibility modifiers for better encapsulation
